### 2020-06-28 #44 @luksurious
[Re] Faster Teaching via POMDP Planning (partial replication)
**Original article:** Faster Teaching via POMDP Planning by Rafferty et. al (2016) , https://www.onlinelibrary.wiley.com/doi/full/10.1111/cogs.12290

**Scientific domain:** Cognitive Science
**Suggested editor:** -

(Sorry, I had first put the wrong link to the paper, it is now working well)
