### 2020-06-29 #44 @rougier
@gdetor @koustuvsinha Can you edit this (regular) submission (Faster Teaching via POMDP Planning)?
