### 2020-07-06 #44 @rougier
@koustuvsinha Can you edit this (regular) submission (Faster Teaching via POMDP Planning)?
