### 2020-08-06 #44 @rougier
@koustuvsinha No problem. Actually the request is for editing, meaning you just need to assign 2 reviewers (from the board or from elsewhere). Or you can edit and review and find another reviewer.
