### 2020-08-08 #44 @benureau
> Hi @benureau 👋 would you be interested to review this article?

Dear @koustuvsinha, I have personally worked with Aurélien Nioche, the second author,  in the same lab, during my last position. I don't think I can review this paper.
