### 2020-08-11 #44 @koustuvsinha
Hi @bengioe, as we discussed in email, would you be interested to review this article? You can find the [reviewer guidelines](https://rescience.github.io/edit/) here. Many thanks!
