### 2020-08-11 #44 @bengioe
Yes, I can review this paper.
