### 2020-08-11 #44 @koustuvsinha
Thanks so much @bengioe! Additionally, it would be great if you can comment in [this thread](https://github.com/ReScience/ReScience/issues/27) with your credentials, so that we can onboard you as a reviewer. @rougier, can you confirm that is the process to onboard external reviewers?
