### 2020-08-17 #44 @rougier
Yes, it's not mandatory at all but it's better. I'll add @bengioe to the board.
