### 2020-08-26 #44 @bengioe
Hello, here is my review. As this is my first review here, I'm happy to extend it if more details are necessary.

**Original Paper** The original paper being reproduced is a 2016 paper using the Partially Observable Markov Decision Process framework in conjunction with heuristic student models to learn teaching policies. The original paper simulates students through their heuristic models, and also performs human trials to validate their approach.

Specifically, 3 student models are proposed, a memoryless model, a short-term memory model, and a continuous belief model (estimated with particle filters). In such a model, a weighing of the enumeration of all possible concepts consists of the state space, while the action space consists of the 3 possible actions for the teacher (show an example, quiz, quiz with feedback).

The original paper finds that using appropriate student models to estimate the optimal policy can improve teaching.

**Reproduction** This study reproduces the simulated experiments of the original paper. While most findings are not significantly different, the results that are seem more consistent with expectations about the algorithms that are used. In addition, this study highlights some errors of the original paper and explicitly provides many details useful to reproduction.

- **audience** - I estimate that an audience with a generic CS background should be able to understand this study. The details of the POMDP framework as well as those of the individual learned models are explained clearly.
- **level of detail** - all details needed to reproduce the simulations and reimplement the proposed algorithms are present. I appreciate in particular Table 5, which for me clarified the scope of the Number Game of the original paper.
- **discussion** - the study contains an interesting discussion of the reproduction of the results, the differences found, as well as the choices made to create this experimental setup (such as the relative cost of actions). In particular it weakens some of the conclusions of the original paper, leaving place for improvement and additional research on this setting.
- **writing** - The writing was clear throughout, I have made a few suggestions below.

Code comments:
- I was able to run the code successfully. The code is fairly clean and self-explanatory.
- I would suggest adding a `mkdir -p data` command in the `.sh` files to avoid failure from a fresh clone of the repository.
- Some code seems to use `random` instead of `numpy.random`. While I see that you seed both modules, it would be preferable to only use one. (for example, prior to numpy1.17 both `random` and `numpy.random` use the same PRNG algorithm, MT19937, so seeding both with the same seed should produce the exact same sequence of numbers, which isn't desirable).

Additional comments:
- Section 2.1, you introduce Q(s,a) in (2) and only define it later with (3). It may worth very shortly introducing Q right after (2), or even to more generally explain what action-value functions are.
- Sec 3.2, 3.3, The noise/cost parameters (Table 1-4) appear quite specific, it may be worth mentioning the noises come from Corbett & Anderson, 1995 (I think? This is what I understood from the original paper, which you cite as their source), and that the cost come from control human experiments and represent seconds.
- Sec 3.6, "Similar to the particle filter in the continuous model where, the belief [are reset]", did you mean "Similarly, for the particle filter in the continuous model, the beliefs [are reset]"
- You end the paper with this: "Through this replication, we hope to facilitate research in this direction." I think it would be nice to argue why this is valuable; e.g. is it easy to write down analytical learned models? Do modern computational capacities allow to use the full potential of POMDPs?

Note on style:
- in the original paper, both students and teachers are referred to as "she" and "her", whereas in this paper there is a mix of "he/his" and "their". Note that in English using "they" is a valid [gender-neutral singular pronoun](https://en.wikipedia.org/wiki/Singular_they). If you wish to, I'd suggest either using "she/her" in honor of the original paper or "they/their" for consistency.
