### 2020-08-31 #44 @koustuvsinha
@amyzhang has accepted to review this paper! (as discussed in email). Thanks a lot! :) You can find the [reviewer guidelines](https://rescience.github.io/edit/) here.
