### 2020-09-06 777c0b2
review code comments: replace random with np.random (and upgrade to Numpy >1.17), add data dir creation in shell scripts
