+++ number_game_mul7_exp.sh
@@ -2,6 +2,9 @@
 
 COMMON_ARGS='--number_concept mul7 --plan_pre_steps 20 --teaching_phase_actions 5 '
 
+# Ensure target folder of log files exists
+mkdir -p data
+
 # Memoryless planning
 python -u main.py $COMMON_ARGS --sim_model memoryless --plan_pre_horizon 2 --plan_online_horizon 2 --plan_online_samples 6 8 -- memoryless number_game > "data/ng_mul7_mless-mless.log"
 python -u main.py $COMMON_ARGS --sim_model discrete --plan_pre_horizon 2 --plan_online_horizon 2 --plan_online_samples 6 8 --plan_load_actions data/actions.pickle -- memoryless number_game > "data/ng_mul7_mless-discrete.log"
