### 2020-09-08 #44 @luksurious
@bengioe Thank you for your review.

We have addressed your comments.

* The code uses only `np.random` now (with the upgraded NG system in 1.17+), and the setup is self-contained (also removed some test output I left there)
* The paper is updated to address your points.
  * The Q function is correctly introduced, belief reset is better explained, and the conclusion is extended.
  * Re Sec 3.2, 3.3, noise parameters: They were also fitted from the control experiments with humans in the original study. This is explained in detail in the supplementary material of the original paper (for reference: [Supplementary material](https://onlinelibrary.wiley.com/action/downloadSupplement?doi=10.1111%2Fcogs.12290&file=cogs12290-sup-0001-SuppInfo.pdf)). We added a note to clarify this.
  * We employed now `they` as a gender-neutral pronoun

Let me know if there are additional points to address.

CC @AurelienNioche
