### 2020-10-08 #44 @amyzhang
**Replication.** The original paper uses POMDPs to formulate a teacher-student setting. The selection of the next teaching activity is a planning problem, where the teacher maintains a belief state of the student. There are three models proposed, and evaluated on two tasks.
The three models consist of a memoryless model, a discrete model with memory, explicitly keeping a history of the past m actions, and a continuous model with implicit information about the entire history.
The two tasks are a simple letter arithmetic task with the goal of finding the correct mapping between a set of letters and numbers, and a number game where the students learn a target number concept for general numbers, such as odd numbers or numbers within some range.
Reproducibility of the replication. The authors were able to reproduce the results for the first task. However, in the second task their results differ. Specifically, the authors found that they could train policies with the three methods that perform better than random, but not necessarily better than baselines. Further, there were failure modes for certain policies paired with certain learner models.

**Clarity of code and instructions.** The instructions were clear, and there were separate scripts to run for each task, which made running the results very simple. I was able to set up my environment and run the scripts on the first try using the README in the code.

**Clarity and completeness of the accompanying article.** The article is clear and well written, with a general description of the contributions and evaluations in the original paper, and the findings in the beginning. I have a high-level suggestion to make clear what components and details in the methods section are taken from the original paper, and what was unclear from the original paper and therefore required design decisions on the part of the authors. When reading the paper, it is unclear if any liberties needed to be taken from the original method.
I appreciated the analysis in the Experiments section for Task 2, where the authors lay out the differences in version of the task and their design choice. The discussion sections with analysis on the potential failure modes in the method and experiments was also very useful and answered some of my earlier questions as to what was the original method and what required changes on the authors' part in order to replicate the results. Some components should go in the methods section to highlight the contributions and possible deviations from the original method rather than getting buried at the end. Several of the conjectures for improvement in the discussion section are also very interesting, and it would be nice to see if they are backed up empirically -- but perhaps that is outside the scope of this work.

Other than the high-level reshuffling of some paragraphs in the Discussion section into the Methods section and added clarity into what was explicitly described in the original paper and what was not, I don't have any suggestions for improvements. I found the explanation of the method and tasks clear, and the results, analysis, and discussion insightful. Great job!
