### 2020-10-12 #44 @luksurious
@amyzhang  Thank you for your review.

We verified all questions and unclear elements with the original authors.
I have added a subsection at the end of Methods to make this more explicit.

Except for the remark about the belief update, which is mentioned now at the end of the methods section, I did not find any other remarks in the discussion that refer to changes compared to the original description.
Please let me know if there are particular segments remaining that should be addressed earlier in the methods section.

Thanks!
