### 2020-11-03 #44 @koustuvsinha
Hi @luksurious, thanks for your remarks, and @amyzhang and @bengioe thanks for your valuable reviews. I believe the submission has acknowledged the proposed changes. Unless the reviewers feel strongly about the rebuttal, I vote for acceptance of the paper.
