### 2020-11-03 #44 @koustuvsinha
Great!! Congrats @luksurious, your paper is now accepted to ReScience journal! 🎉 I'll follow up soon with a PR to your repository with the correct article numbers.
