### 2020-11-03 #44 @koustuvsinha
@luksurious I couldn't find the ReScience template tex sources in your repository. Can you add them such that I can compile on my end?
