### 2020-11-03 #44 @luksurious
@koustuvsinha Ah sorry, I was working with it in Overleaf, so I synced it now to a new repository. You can find it here
https://github.com/luksurious/faster-teaching-paper
