### 2020-11-03 #44 @koustuvsinha
Article published on Zenodo and will be available shortly on ReScience website!

- Article PR: https://github.com/ReScience/articles/pull/20
- Website PR: https://github.com/ReScience/rescience.github.io/pull/96

This concludes the reviewing and editing process of this paper.
